fix bug.. like the one where konqy would start after clicking a radical in kanji view.. clean up things like the gjiten heuristic thing.. add a tooltip and fix bug in radical thing.. add new search action..
svn path=/trunk/kdeedu/kiten/; revision=314678

+++ learn.cpp
@@ -697,13 +697,13 @@ QString Learn::randomMeaning(QStringList &oldMeanings)
 			switch (guessOn)
 			{
 				case 1:
-				meaning = shortenString(Dict::prettyMeaning((*list.at(rand)).meanings()));
+				meaning = shortenString(Dict::prettyMeaning((*list.at(static_cast<unsigned int>(rand))).meanings()));
 				break;
 				case 2:
-				meaning = Dict::prettyKanjiReading((*list.at(rand)).readings());
+				meaning = Dict::prettyKanjiReading((*list.at(static_cast<unsigned int>(rand))).readings());
 				break;
 				case 0:
-				meaning = (*list.at(rand)).kanji();
+				meaning = (*list.at(static_cast<unsigned int>(rand))).kanji();
 			}
 		}
 		else
